+++ lib/npm/cache.ex
@@ -40,6 +40,8 @@ defmodule NPM.Cache do
     if cached?(name, version) do
       {:ok, dest}
     else
+      NPM.Security.RegistryPolicy.validate_url!(tarball_url)
+
       case NPM.Tarball.fetch_and_extract(tarball_url, integrity, dest) do
         {:ok, _count} ->
           {:ok, dest}